Extract /apikey/... to separate module
I had to extract some shared stuff into other modules in order to
support doing this.
