+++ app/api/routes/logger.py
@@ -1,0 +1,3 @@
+from app import utils as utils
+
+logger = utils.setup_logger('routes_logger')
